fix: Create a single session for testing
